requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
PyOpenGL==3.1.7
pygame==2.5.2
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0